code/requirements.txt
torch>=2.0.0
torchvision
timm
pillow
tqdm
scikit-learn
matplotlib
numpy
